// libs/domain/src/purchase-and-orders/model/goods-delivery.model.ts
import { GoodsDeliveryTypeEnum } from '../enum/goods-delivery-type.enum';
import { GoodsDeliveryStateEnum } from '../enum/goods-delivery-state.enum';
import { GoodsDeliveryBaseDocumentTypeEnum } from '../enum/goods-delivery-base-document-type.enum';
import { GoodsDeliveryItemModel } from './goods-delivery-item.model';
import { Pagination } from '@view/lib/models';
import { ClSelectButtonOption } from '@sadad/component-lib/src/models';



export interface GoodsDeliveryModel extends Pagination {
  id?: number;
  deliveryType?: GoodsDeliveryTypeEnum;
  deliveryTypeString?: string;
  autoGeneratedCode?: number;
  state?: GoodsDeliveryStateEnum;
  nextState?: GoodsDeliveryStateEnum;
  stateString?: string;
  date?: Date;
  fromDate?: Date;
  toDate?: Date;
  datePersian?: string;
  delivererId?: number;
  delivererLabel?: string;
  delivererType?: ClSelectButtonOption & {value?: 'prs-corps' | 'inventories' | 'organizations'};
  receiverId?: number;
  receiverLabel?: string;
  receiverType?: ClSelectButtonOption & {value?: 'prs-corps' | 'inventories' | 'organizations'};
  fiscalYearId?: number;
  baseDocumentType?: GoodsDeliveryBaseDocumentTypeEnum;
  baseDocumentTypeString?: string;
  baseDocumentNumber?: number;
  baseDocumentDate?: Date;
  baseDocumentDatePersian?: string;
  description?: string;
  deliveryItemsList?: GoodsDeliveryItemModel[];
  ifDeliveryItemsLoaded?: boolean;

  sendDate?: Date; // تاریخ ارسال
  receiverFullName?: string; // تحویل گیرنده
  sourceAdditionalComments?: string; // توضیحات ارسال

  receiveDate?: Date; // تاریخ دریافت
  delivererFullName?: string; // تحویل دهنده
  destinationAdditionalComments?: string; // توضیحات دریافت
}


export type GoodsDeliveryModelFilter = keyof GoodsDeliveryModel;